# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: EV Charging Platform API
  version: "1.0"
  description: >
    Core REST surface for driver, fleet, and operator clients. Real-time
    session telemetry (live kWh/cost) is delivered over a separate
    WebSocket channel (/ws/sessions/{session_id}) fed by the OCPP Central
    System — not polled over REST.
servers:
  - url: https://api.evplatform.com/v1
security:
  - bearerAuth: []
paths:
  /stations/search:
    get:
      summary: Search nearby stations
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lng
          in: query
          required: true
          schema: { type: number }
        - name: radius_km
          in: query
          schema: { type: number, default: 10 }
        - name: connector_type
          in: query
          schema: { type: string, enum: [CCS2, CHAdeMO, TYPE2, NACS] }
        - name: min_power_kw
          in: query
          schema: { type: number }
        - name: available_only
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Matching stations with live connector status
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StationSummary" }
  /stations/{stationId}:
    get:
      summary: Station detail — connectors, pricing, reliability scores
      parameters:
        - { name: stationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Full station detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StationDetail" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StationSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        lat: { type: number }
        lng: { type: number }
        distance_km: { type: number }
        min_price_per_kwh: { type: number }
        connectors_available: { type: integer }
        connectors_total: { type: integer }
        reliability_score: { type: number, description: "0-100, per-station rollup" }
    StationDetail:
      allOf:
        - $ref: "#/components/schemas/StationSummary"
        - type: object
          properties:
            address: { type: string }
            connectors:
              type: array
              items: { $ref: "#/components/schemas/Connector" }
            tariff: { $ref: "#/components/schemas/Tariff" }
    Connector:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [CCS2, CHAdeMO, TYPE2, NACS] }
        power_kw: { type: number }
        status: { type: string, enum: [available, occupied, faulted, reserved] }
        reliability_score: { type: number }
...
    Tariff:
      type: object
      properties:
        pricing_model: { type: string, enum: [per_kwh, per_minute, flat] }
        rate: { type: number }
        time_of_day_rules:
          type: array
          items:
            type: object
            properties:
              start: { type: string }
              end: { type: string }
              multiplier: { type: number }
...